+++ chain/params.go
@@ -1,5 +1,5 @@
-// Copyright (c) 2019 KIDTSUNAMI
-// Author: alex@kidtsunami.com
+// Copyright (c) 2020 Blockwatch Data Inc.
+// Author: alex@blockwatch.cc
 
 package chain
 
@@ -33,7 +33,7 @@ type Params struct {
 	EndorsersPerBlock            int              `json:"endorsers_per_block"`
 	HardGasLimitPerOperation     int64            `json:"hard_gas_limit_per_operation"`
 	HardGasLimitPerBlock         int64            `json:"hard_gas_limit_per_block"`
-	ProofOfWorkThreshold         uint64           `json:"proof_of_work_threshold"`
+	ProofOfWorkThreshold         int64            `json:"proof_of_work_threshold"`
 	ProofOfWorkNonceSize         int              `json:"proof_of_work_nonce_size"`
 	TokensPerRoll                int64            `json:"tokens_per_roll"`
 	MichelsonMaximumTypeSize     int              `json:"michelson_maximum_type_size"`
@@ -130,4 +130,12 @@ func (p *Params) VotingStartCycleFromHeight(height int64) int64 {
 	currentCycle := p.CycleFromHeight(height)
 	offset := height % p.BlocksPerVotingPeriod
 	return currentCycle - offset/p.BlocksPerCycle
+}
+
+func (p *Params) IsVoteStart(height int64) bool {
+	return height > 0 && (height-1)%p.BlocksPerVotingPeriod == 0
+}
+
+func (p *Params) IsVoteEnd(height int64) bool {
+	return height > 0 && height%p.BlocksPerVotingPeriod == 0
 }