+++ micheline/bigmap.go
@@ -1,5 +1,5 @@
-// Copyright (c) 2019 KIDTSUNAMI
-// Author: alex@kidtsunami.com
+// Copyright (c) 2020 Blockwatch Data Inc.
+// Author: alex@blockwatch.cc
 
 package micheline
 
@@ -109,21 +109,21 @@ func (e *BigMapDiffElem) Encoding() PrimType {
 	return PrimTypeFromTypeCode(e.KeyType)
 }
 
-// func (e *BigMapDiffElem) Key() *BigMapKey {
-// 	k, err := NewBigMapKey(e.KeyType, e.IntKey, e.StringKey, e.BytesKey, e.Value)
-// 	if err != nil {
-// 		log.Error(err)
-// 	}
-// 	return k
-// }
+func (e *BigMapDiffElem) Key() *BigMapKey {
+	k, err := NewBigMapKey(e.KeyType, e.IntKey, e.StringKey, e.BytesKey, e.Value)
+	if err != nil {
+		log.Error(err)
+	}
+	return k
+}
 
-// func (e *BigMapDiffElem) KeyAs(typ OpCode) *BigMapKey {
-// 	k, err := NewBigMapKey(typ, e.IntKey, e.StringKey, e.BytesKey, e.Value)
-// 	if err != nil {
-// 		log.Error(err)
-// 	}
-// 	return k
-// }
+func (e *BigMapDiffElem) KeyAs(typ OpCode) *BigMapKey {
+	k, err := NewBigMapKey(typ, e.IntKey, e.StringKey, e.BytesKey, e.Value)
+	if err != nil {
+		log.Error(err)
+	}
+	return k
+}
 
 func (e *BigMapDiffElem) KeyString() string {
 	switch true {