+++ rpc/mempool.go
@@ -1,6 +1,6 @@
 // Copyright (c) 2018 ECAD Labs Inc. MIT License
-// Copyright (c) 2019 KIDTSUNAMI
-// Author: alex@kidtsunami.com
+// Copyright (c) 2020 Blockwatch Data Inc.
+// Author: alex@blockwatch.cc
 
 package rpc
 
@@ -20,7 +20,7 @@ type MempoolOperations struct {
 // GetMempoolPendingOperations returns mempool pending operations
 func (c *Client) GetMempoolPendingOperations(ctx context.Context) (*MempoolOperations, error) {
 	var ops MempoolOperations
-	if err := c.Get(ctx, "/chains/"+c.ChainID+"/mempool/pending_operations", &ops); err != nil {
+	if err := c.Get(ctx, "chains/"+c.ChainID+"/mempool/pending_operations", &ops); err != nil {
 		return nil, err
 	}
 	return &ops, nil