+++ server/contract.go
@@ -1,5 +1,5 @@
-// Copyright (c) 2019 KIDTSUNAMI
-// Author: alex@kidtsunami.com
+// Copyright (c) 2020 Blockwatch Data Inc.
+// Author: alex@blockwatch.cc
 
 package server
 
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"time"
 
-	"blockwatch.cc/packdb/vec"
+	"blockwatch.cc/packdb/pack"
 	"blockwatch.cc/tzindex/chain"
 	"blockwatch.cc/tzindex/etl/index"
 	"blockwatch.cc/tzindex/etl/model"
@@ -22,48 +22,45 @@ func init() {
 var _ RESTful = (*ExplorerContract)(nil)
 
 type ExplorerContract struct {
-	Address            string            `json:"address"`
-	Manager            string            `json:"manager"`
-	Delegate           string            `json:"delegate"`
-	Height             int64             `json:"height"`
-	Fee                float64           `json:"fee"`
-	GasLimit           int64             `json:"gas_limit"`
-	GasUsed            int64             `json:"gas_used"`
-	GasPrice           float64           `json:"gas_price"`
-	StorageLimit       int64             `json:"storage_limit"`
-	StorageSize        int64             `json:"storage_size"`
-	StoragePaid        int64             `json:"storage_paid"`
-	Script             *micheline.Script `json:"script,omitempty"` // DEPRECATED
-	IsFunded           bool              `json:"is_funded"`
-	IsVesting          bool              `json:"is_vesting"`
-	IsSpendable        bool              `json:"is_spendable"`
-	IsDelegatable      bool              `json:"is_delegatable"`
-	IsDelegated        bool              `json:"is_delegated"`
-	FirstIn            int64             `json:"first_in"`
-	FirstOut           int64             `json:"first_out"`
-	LastIn             int64             `json:"last_in"`
-	LastOut            int64             `json:"last_out"`
-	FirstSeen          int64             `json:"first_seen"`
-	LastSeen           int64             `json:"last_seen"`
-	DelegatedSince     int64             `json:"delegated_since"`
-	FirstInTime        time.Time         `json:"first_in_time"`
-	FirstOutTime       time.Time         `json:"first_out_time"`
-	LastInTime         time.Time         `json:"last_in_time"`
-	LastOutTime        time.Time         `json:"last_out_time"`
-	FirstSeenTime      time.Time         `json:"first_seen_time"`
-	LastSeenTime       time.Time         `json:"last_seen_time"`
-	DelegatedSinceTime time.Time         `json:"delegated_since_time"`
-	NOps               int               `json:"n_ops"`
-	NOpsFailed         int               `json:"n_ops_failed"`
-	NTx                int               `json:"n_tx"`
-	NDelegation        int               `json:"n_delegation"`
-	NOrigination       int               `json:"n_origination"`
-	TokenGenMin        int64             `json:"token_gen_min"`
-	TokenGenMax        int64             `json:"token_gen_max"`
-	DelegateAcc        *ExplorerAccount  `json:"delegate_account,omitempty"` // DEPRECATED
-	ManagerAcc         *ExplorerAccount  `json:"manager_account,omitempty"`  // DEPRECATED
-	Ops                *[]*ExplorerOp    `json:"ops,omitempty"`              // DEPRECATED
-	expires            time.Time         `json:"-"`
+	Address            string    `json:"address"`
+	Manager            string    `json:"manager"`
+	Delegate           string    `json:"delegate"`
+	Height             int64     `json:"height"`
+	Fee                float64   `json:"fee"`
+	GasLimit           int64     `json:"gas_limit"`
+	GasUsed            int64     `json:"gas_used"`
+	GasPrice           float64   `json:"gas_price"`
+	StorageLimit       int64     `json:"storage_limit"`
+	StorageSize        int64     `json:"storage_size"`
+	StoragePaid        int64     `json:"storage_paid"`
+	IsFunded           bool      `json:"is_funded"`
+	IsVesting          bool      `json:"is_vesting"`
+	IsSpendable        bool      `json:"is_spendable"`
+	IsDelegatable      bool      `json:"is_delegatable"`
+	IsDelegated        bool      `json:"is_delegated"`
+	FirstIn            int64     `json:"first_in"`
+	FirstOut           int64     `json:"first_out"`
+	LastIn             int64     `json:"last_in"`
+	LastOut            int64     `json:"last_out"`
+	FirstSeen          int64     `json:"first_seen"`
+	LastSeen           int64     `json:"last_seen"`
+	DelegatedSince     int64     `json:"delegated_since"`
+	FirstInTime        time.Time `json:"first_in_time"`
+	FirstOutTime       time.Time `json:"first_out_time"`
+	LastInTime         time.Time `json:"last_in_time"`
+	LastOutTime        time.Time `json:"last_out_time"`
+	FirstSeenTime      time.Time `json:"first_seen_time"`
+	LastSeenTime       time.Time `json:"last_seen_time"`
+	DelegatedSinceTime time.Time `json:"delegated_since_time"`
+	NOps               int       `json:"n_ops"`
+	NOpsFailed         int       `json:"n_ops_failed"`
+	NTx                int       `json:"n_tx"`
+	NDelegation        int       `json:"n_delegation"`
+	NOrigination       int       `json:"n_origination"`
+	TokenGenMin        int64     `json:"token_gen_min"`
+	TokenGenMax        int64     `json:"token_gen_max"`
+	BigMapIds          []int64   `json:"bigmap_ids"`
+	expires            time.Time `json:"-"`
 }
 
 func NewExplorerContract(ctx *ApiContext, c *model.Contract, a *model.Account, p *chain.Params, details bool) *ExplorerContract {
@@ -99,18 +96,6 @@ func NewExplorerContract(ctx *ApiContext, c *model.Contract, a *model.Account, p
 		expires:        ctx.Now.Add(p.TimeBetweenBlocks[0]),
 	}
 
-	// need manager address hash
-	var mgrHash []byte
-	if mgr, err := ctx.Indexer.LookupAccountId(ctx, c.ManagerId); err == nil {
-		mgrHash = mgr.Address().Bytes()
-	}
-	tip := ctx.Crawler.Tip()
-	if sc, err := c.LoadScript(tip, tip.BestHeight, mgrHash); err != nil {
-		log.Errorf("explorer contract: unmarshal script: %v", err)
-	} else {
-		cc.Script = sc
-	}
-
 	// resolve block times
 	cc.FirstInTime = ctx.Indexer.BlockTime(ctx.Context, a.FirstIn)
 	cc.FirstOutTime = ctx.Indexer.BlockTime(ctx.Context, a.FirstOut)
@@ -120,31 +105,14 @@ func NewExplorerContract(ctx *ApiContext, c *model.Contract, a *model.Account, p
 	cc.LastSeenTime = ctx.Indexer.BlockTime(ctx.Context, a.LastSeen)
 	cc.DelegatedSinceTime = ctx.Indexer.BlockTime(ctx.Context, a.DelegatedSince)
 
-	// DEPRECATED
-	if details {
-		// load related accounts from id
-		xc, err := ctx.Indexer.LookupAccountIds(ctx.Context,
-			vec.UniqueUint64Slice([]uint64{
-				a.ManagerId.Value(),
-				a.DelegateId.Value(),
-			}))
-		if err != nil {
-			log.Errorf("explorer contract: cannot resolve related accounts: %v", err)
-		}
-		for _, xcc := range xc {
-			if xcc.RowId == a.ManagerId {
-				cc.Manager = xcc.String()
-				cc.ManagerAcc = NewExplorerAccount(ctx, xcc, p, false)
-			}
-			if xcc.RowId == a.DelegateId {
-				cc.Delegate = xcc.String()
-				cc.DelegateAcc = NewExplorerAccount(ctx, xcc, p, false)
-			}
-		}
-	} else {
-		cc.Manager = lookupAddress(ctx, a.ManagerId).String()
-		cc.Delegate = lookupAddress(ctx, a.DelegateId).String()
+	var err error
+	cc.BigMapIds, err = ctx.Indexer.ListContractBigMapIds(ctx.Context, a.RowId)
+	if err != nil {
+		log.Errorf("explorer contract: cannot load bigmap ids: %v", err)
 	}
+
+	cc.Manager = lookupAddress(ctx, a.ManagerId).String()
+	cc.Delegate = lookupAddress(ctx, a.DelegateId).String()
 
 	return cc
 }
@@ -172,13 +140,81 @@ func (b ExplorerContract) RegisterDirectRoutes(r *mux.Router) error {
 
 func (b ExplorerContract) RegisterRoutes(r *mux.Router) error {
 	r.HandleFunc("/{ident}", C(ReadContract)).Methods("GET").Name("contract")
-	r.HandleFunc("/{ident}/op", C(ReadContractOps)).Methods("GET") // DEPRECATED
+	r.HandleFunc("/{ident}/calls", C(ReadContractCalls)).Methods("GET")
+	r.HandleFunc("/{ident}/manager", C(ReadContractManager)).Methods("GET")
+	r.HandleFunc("/{ident}/script", C(ReadContractScript)).Methods("GET")
+	r.HandleFunc("/{ident}/storage", C(ReadContractStorage)).Methods("GET")
 	return nil
 
 }
 
 type ContractRequest struct {
 	ExplorerListRequest // offset, limit
+
+	Order  pack.OrderType `schema:"order"`
+	Block  string         `schema:"block"`  // height or hash for time-lock
+	Since  string         `schema:"since"`  // block hash or height for updates
+	Unpack bool           `schema:"unpack"` // unpack packed key/values
+	Prim   bool           `schema:"prim"`   // for prim/value rendering
+
+	// decoded values
+	BlockHeight int64           `schema:"-"`
+	BlockHash   chain.BlockHash `schema:"-"`
+	SinceHeight int64           `schema:"-"`
+	SinceHash   chain.BlockHash `schema:"-"`
+}
+
+func (r *ContractRequest) WithPrim() bool {
+	return r != nil && r.Prim
+}
+
+func (r *ContractRequest) WithUnpack() bool {
+	return r != nil && r.Unpack
+}
+
+func (r *ContractRequest) ParseBlockIdent(ctx *ApiContext) {
+	if len(r.Block) > 0 {
+		b, err := ctx.Indexer.LookupBlock(ctx.Context, r.Block)
+		if err != nil {
+			switch err {
+			case index.ErrNoBlockEntry:
+				panic(ENotFound(EC_RESOURCE_NOTFOUND, "no such block", err))
+			case index.ErrInvalidBlockHeight:
+				panic(EBadRequest(EC_RESOURCE_ID_MALFORMED, "invalid block height", err))
+			case index.ErrInvalidBlockHash:
+				panic(EBadRequest(EC_RESOURCE_ID_MALFORMED, "invalid block hash", err))
+			default:
+				panic(EInternal(EC_DATABASE, err.Error(), nil))
+			}
+		}
+		// make sure block is not orphaned
+		if b.IsOrphan {
+			panic(EConflict(EC_RESOURCE_STATE_UNEXPECTED, "block is orphaned", nil))
+		}
+		r.BlockHeight = b.Height
+		r.BlockHash = b.Hash.Clone()
+	}
+	if len(r.Since) > 0 {
+		b, err := ctx.Indexer.LookupBlock(ctx.Context, r.Since)
+		if err != nil {
+			switch err {
+			case index.ErrNoBlockEntry:
+				panic(ENotFound(EC_RESOURCE_NOTFOUND, "no such block", err))
+			case index.ErrInvalidBlockHeight:
+				panic(EBadRequest(EC_RESOURCE_ID_MALFORMED, "invalid block height", err))
+			case index.ErrInvalidBlockHash:
+				panic(EBadRequest(EC_RESOURCE_ID_MALFORMED, "invalid block hash", err))
+			default:
+				panic(EInternal(EC_DATABASE, err.Error(), nil))
+			}
+		}
+		// make sure block is not orphaned
+		if b.IsOrphan {
+			panic(EConflict(EC_RESOURCE_STATE_UNEXPECTED, "block is orphaned", nil))
+		}
+		r.SinceHeight = b.Height
+		r.SinceHash = b.Hash.Clone()
+	}
 }
 
 func loadContract(ctx *ApiContext) *model.Contract {
@@ -216,40 +252,197 @@ func ReadContract(ctx *ApiContext) (interface{}, int) {
 	return NewExplorerContract(ctx, cc, acc, ctx.Crawler.ParamsByHeight(-1), true), http.StatusOK
 }
 
-func ReadContractOps(ctx *ApiContext) (interface{}, int) {
-	args := &ExplorerOpsRequest{}
+// list incoming transaction with data
+func ReadContractCalls(ctx *ApiContext) (interface{}, int) {
+	args := &ContractRequest{}
 	ctx.ParseRequestArgs(args)
 	cc := loadContract(ctx)
 	acc, err := ctx.Indexer.LookupAccountId(ctx, cc.AccountId)
 	if err != nil {
 		switch err {
-		case index.ErrNoContractEntry:
+		case index.ErrNoAccountEntry:
 			panic(ENotFound(EC_RESOURCE_NOTFOUND, "no such contract", err))
 		default:
 			panic(EInternal(EC_DATABASE, err.Error(), nil))
 		}
 	}
+	args.ParseBlockIdent(ctx)
 	params := ctx.Crawler.ParamsByHeight(-1)
-	c := NewExplorerContract(ctx, cc, acc, params, false)
-	ops, err := ctx.Indexer.ListAccountOps(
+	ops, err := ctx.Indexer.ListContractCalls(
 		ctx,
 		acc.RowId,
-		args.Type,
-		0,
-		0,
+		args.SinceHeight+1,
+		args.BlockHeight, // since, until (optional)
 		args.Offset,
-		ctx.Cfg.ClampExplore(args.Limit),
-		0,
+		ctx.Cfg.ClampExplore(args.Limit), // offset, limit (optional)
+		args.Order,
 	)
 	if err != nil {
-		panic(EInternal(EC_DATABASE, "cannot read contract operations", err))
+		panic(EInternal(EC_DATABASE, "cannot read contract calls", err))
 	}
 
+	// we reuse explorer ops here
 	// FIXME: collect account and op lookup into only two queries
 	eops := make([]*ExplorerOp, len(ops))
 	for i, v := range ops {
-		eops[i] = NewExplorerOp(ctx, v, nil, cc, params, nil)
+		eops[i] = NewExplorerOp(ctx, v, nil, cc, params, args)
 	}
-	c.Ops = &eops
-	return c, http.StatusOK
+	return eops, http.StatusOK
+}
+
+func ReadContractManager(ctx *ApiContext) (interface{}, int) {
+	cc := loadContract(ctx)
+	if cc.ManagerId == 0 {
+		panic(ENotFound(EC_RESOURCE_NOTFOUND, "no manager for this contract", nil))
+	}
+	acc, err := ctx.Indexer.LookupAccountId(ctx, cc.ManagerId)
+	if err != nil {
+		switch err {
+		case index.ErrNoAccountEntry:
+			panic(ENotFound(EC_RESOURCE_NOTFOUND, "no such account", err))
+		default:
+			panic(EInternal(EC_DATABASE, err.Error(), nil))
+		}
+	}
+	params := ctx.Crawler.ParamsByHeight(-1)
+	return NewExplorerAccount(ctx, acc, params, false), http.StatusOK
+}
+
+type ExplorerScript struct {
+	Script      *micheline.Script     `json:"script,omitempty"`
+	Type        micheline.BigMapType  `json:"storage_type"`
+	Entrypoints micheline.Entrypoints `json:"entrypoints"`
+	modified    time.Time             `json:"-"`
+}
+
+func (s ExplorerScript) LastModified() time.Time { return s.modified }
+func (s ExplorerScript) Expires() time.Time      { return time.Time{} }
+
+var _ Resource = (*ExplorerScript)(nil)
+
+func ReadContractScript(ctx *ApiContext) (interface{}, int) {
+	args := &ContractRequest{}
+	ctx.ParseRequestArgs(args)
+	cc := loadContract(ctx)
+	args.ParseBlockIdent(ctx)
+
+	// unmarshal and optionally migrate script
+	if args.BlockHeight == 0 {
+		args.BlockHeight = ctx.Crawler.Height()
+	}
+
+	// need manager hash
+	var mgrHash []byte
+	if mgr, err := ctx.Indexer.LookupAccountId(ctx, cc.ManagerId); err == nil {
+		mgrHash = mgr.Address().Bytes()
+	}
+	script, err := cc.LoadScript(ctx.Crawler.Tip(), args.BlockHeight, mgrHash)
+	if err != nil {
+		panic(EInternal(EC_SERVER, "script unmarshal failed", err))
+	}
+
+	ep, err := script.Entrypoints(args.WithPrim())
+	if err != nil {
+		panic(EInternal(EC_SERVER, "script entrypoint parsing failed", err))
+	}
+
+	resp := &ExplorerScript{
+		Script:      script,
+		Type:        script.StorageType(),
+		Entrypoints: ep,
+		modified:    ctx.Indexer.BlockTime(ctx.Context, cc.Height),
+	}
+	if !args.WithPrim() {
+		resp.Script = nil
+	}
+
+	return resp, http.StatusOK
+}
+
+func ReadContractStorage(ctx *ApiContext) (interface{}, int) {
+	args := &ContractRequest{}
+	ctx.ParseRequestArgs(args)
+	cc := loadContract(ctx)
+	args.ParseBlockIdent(ctx)
+
+	if args.BlockHeight > 0 && args.BlockHeight < cc.Height {
+		panic(ENotFound(EC_RESOURCE_NOTFOUND, "empty storage before origination", nil))
+	}
+
+	var mgrHash []byte
+	if mgr, err := ctx.Indexer.LookupAccountId(ctx, cc.ManagerId); err == nil {
+		mgrHash = mgr.Address().Bytes()
+	}
+
+	var (
+		prim   *micheline.Prim
+		typ    *micheline.Prim
+		ts     time.Time
+		height int64
+	)
+
+	// find most recent incoming call
+	op, err := ctx.Indexer.FindLastCall(ctx.Context, cc.AccountId, args.BlockHeight)
+	if err != nil {
+		if err != index.ErrNoOpEntry {
+			panic(EInternal(EC_DATABASE, err.Error(), nil))
+		} else {
+			// unmarshal and optionally migrate script at origination height
+			script, err := cc.LoadScript(ctx.Crawler.Tip(), cc.Height, mgrHash)
+			if err != nil {
+				panic(EInternal(EC_SERVER, "script unmarshal failed", err))
+			}
+			// use storage init from origination, may contain full bigmap definition
+			prim = script.Storage
+			typ = script.Code.Storage.Args[0]
+			ts = ctx.Indexer.BlockTime(ctx.Context, cc.Height)
+			height = cc.Height
+		}
+	} else {
+		// unmarshal latest storage update from op, contains bigmap reference
+		ts = op.Timestamp
+		height = op.Height
+		prim = &micheline.Prim{}
+		if err := prim.UnmarshalBinary(op.Storage); err != nil {
+			log.Errorf("explorer: storage unmarshal in op %s: %v", op.Hash, err)
+		}
+		// unmarshal and optionally migrate script at operation height
+		script, err := cc.LoadScript(ctx.Crawler.Tip(), op.Height, mgrHash)
+		if err != nil {
+			panic(EInternal(EC_SERVER, "script unmarshal failed", err))
+		}
+		typ = script.Code.Storage.Args[0]
+	}
+
+	hash, _ := ctx.Indexer.BlockHashByHeight(ctx.Context, height)
+
+	resp := &ExplorerStorageValue{
+		Meta: ExplorerStorageMeta{
+			Contract: cc.String(),
+			Time:     ts,
+			Height:   height,
+			Block:    hash,
+		},
+		Value: &micheline.BigMapValue{
+			Type:  typ,
+			Value: prim,
+		},
+		modified: ts,
+		expires:  ctx.Crawler.Time().Add(ctx.Crawler.ParamsByHeight(-1).TimeBetweenBlocks[0]),
+	}
+
+	if args.WithPrim() {
+		resp.Prim = prim
+	}
+
+	if args.WithUnpack() && prim.IsPackedAny() {
+		if p, err := prim.UnpackAny(); err == nil {
+			resp.ValueUnpacked = &micheline.BigMapValue{
+				Type:  p.BuildType(),
+				Value: p,
+			}
+		}
+	}
+
+	return resp, http.StatusOK
 }